webkit/data/layout_tests/pending/http/tests/security/cross-frame-access-frames-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://localhost:9000/security/resources/cross-frame-iframe-for-get-test.html from frame with URL http://127.0.0.1:9000/security/cross-frame-access-frames.html. Domains, protocols and ports must match.



----- tests for getting/setting the frames array access and the properties associated with it -----

PASS: targetWindow.frames.length should be '1' and is.
PASS: canGet('targetWindow.frames[0]') should be 'true' and is.
PASS: canGet('targetWindow[0]') should be 'true' and is.
PASS: targetWindow.frames.private should be 'undefined' and is.
